# backend/docs/afnor-contracts/afnor-flow-openapi-v1.3.0.json
{
  "components": {
...
    "requestBodies": {
...
      "FlowSearchRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SearchFlowParams"
            }
          }
        }
      },
...
    },
    "responses": {
      "Error400Response": {
        "content": {
          "application/json": {
            "example": {
              "errorCode": "MISSING_REQUIRED_FIELD",
              "errorMessage": "Invalid request format"
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Error 400 : Bad request."
      },
      "Error401Response": {
        "content": {
          "application/json": {
            "example": {
              "errorCode": "MISSING_TOKEN",
              "errorMessage": null
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Error 401 : Authentication error."
      },
      "Error403Response": {
        "content": {
          "application/json": {
            "example": {
              "errorCode": "FORBIDDEN_ACCESS",
              "errorMessage": null
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Error 403 : Forbidden."
      },
      "Error404Response": {
        "content": {
          "application/json": {
            "example": {
              "errorCode": "MISSING_RESOURCE",
              "errorMessage": "Requested resource is Not Found"
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Error 404 : Resource Not Found."
      },
...
      "Error429Response": {
        "content": {
          "application/json": {
            "example": {
              "errorCode": "TOO_MANY_REQUESTS",
              "errorMessage": "The client emitted too many requests"
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Error 429 : Too many requests."
      },
      "Error500Response": {
        "content": {
          "application/json": {
            "example": {
              "errorCode": "INTERNAL_ERROR",
              "errorMessage": null
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Error 500 : Server Internal Error."
      },
      "Error503Response": {
        "content": {
          "application/json": {
            "example": {
              "errorCode": "RESOURCE_ERROR",
              "errorMessage": null
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Error 503 : Unavailable Resource."
      },
      "FlowGetResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Flow"
            }
          },
          "application/octet-stream": {}
        },
        "description": "OK - Response message once the flow has been downloaded.",
        "headers": {
          "Content-Disposition": {
            "description": "Header specifying attached file name",
            "schema": {
              "pattern": "attachment;filename=*",
              "type": "string"
            }
          }
        }
      },
...
      "FlowSearchResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SearchFlowContent"
            }
          }
        },
        "description": "OK - Response message when returning the results of a search request.\n"
      },
...
    },
    "schemas": {
...
      "CoreFlowInfo": {
        "description": "Core schema to build FlowInfo, FullFlowInfo, Flow schemas",
        "properties": {
          "flowProfile": {
            "$ref": "#/components/schemas/FlowProfile"
          },
          "flowSyntax": {
            "$ref": "#/components/schemas/FlowSyntax"
          },
          "name": {
            "description": "Name of the file",
            "maxLength": 255,
            "type": "string"
          },
          "processingRule": {
            "$ref": "#/components/schemas/ProcessingRule"
          },
          "trackingId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NotOnlyUuid"
              }
            ],
            "description": "The tracking id is an external identifier and is used to track the flow by the sender"
          }
        },
        "required": [
          "flowSyntax",
          "name"
        ],
        "type": "object"
      },
      "Cursor": {
        "description": "Pagination, Opaque cursor for pagination management, acts as a bookmark",
        "example": "eyJjcmVhdGVkX2F0IjoiMjAyNS0wNi0wOVQxMDowMDowMFoiLCJpZCI6MTIzNH0=",
        "type": "string"
      },
      "Error": {
        "description": "Error code",
        "properties": {
          "details": {
            "example": "Error detailed description",
            "type": "string"
          },
          "errorCode": {
            "description": "Short numerical or alphanumerical code that identifies precisely a unique error.",
            "example": "MISSING_REQUIRED_FIELD",
            "type": "string"
          },
          "errorMessage": {
            "description": "Contains information on the error. Not intended to be displayed to an end user. For security reasons, a tradeoff between clarity & security shall be found.",
            "example": "Invalid request format",
            "type": "string"
          },
          "instance": {
            "example": "/chemin/url-de-reference#instance-du-contexte",
            "format": "uri-reference",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "example": "/path/reference-url",
            "format": "uri-reference",
            "type": "string"
          }
        },
        "required": [
          "errorCode"
        ],
        "type": "object"
      },
      "Flow": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CoreFlowInfo"
          },
          {
            "$ref": "#/components/schemas/FullFlowInfoExtension"
          },
          {
            "$ref": "#/components/schemas/FlowExtension"
          }
        ],
        "description": "The properties of a Flow resource",
        "required": [
          "flowProfile",
          "processingRule"
        ]
      },
      "FlowAckStatus": {
        "description": "- `Ok`: the following checks have passed:\n  - Anti virus\n  - Integrity checks\n  - Technical rules checks\n  - Unicity checks\n- `Error`: one of the previous test has failed\n- `Pending` : the flow is not yet integrated\n",
        "enum": [
          "Pending",
          "Ok",
          "Error"
        ],
        "type": "string"
      },
...
      "FlowProfile": {
        "enum": [
          "Basic",
          "CIUS",
          "Extended-CTC-FR",
          "Undefined"
        ],
        "type": "string"
      },
...
      "Limit": {
        "default": 25,
        "description": "Pagination, Maximum number of results that may be returned",
        "maximum": 100,
        "type": "integer"
      },
      "NotOnlyUuid": {
        "description": "Unique identifier supporting UUID but not only, for flexibility purpose",
        "maxLength": 64,
        "type": "string"
      },
...
      "SearchFlowContent": {
        "description": "A set of flows matching criterias, provided into the request",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/SearchFlowFilters"
          },
          "limit": {
            "$ref": "#/components/schemas/Limit"
          },
          "nextCursor": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Cursor"
              }
            ],
            "description": "Pagination management:\n- If present, says where to start for next page, set nextCursor as the cursor in the next request\n- If omitted, says that pagination is finished\n"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Flow"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "SearchFlowFilters": {
        "description": "Filtering criteria, at least one is required",
        "properties": {
          "ackStatus": {
            "$ref": "#/components/schemas/FlowAckStatus"
          },
          "flowDirection": {
            "items": {
              "$ref": "#/components/schemas/FlowDirection"
            },
            "type": "array"
          },
          "flowType": {
            "items": {
              "$ref": "#/components/schemas/FlowType"
            },
            "type": "array"
          },
          "processingRule": {
            "items": {
              "$ref": "#/components/schemas/ProcessingRule"
            },
            "type": "array"
          },
          "trackingId": {
            "$ref": "#/components/schemas/NotOnlyUuid"
          },
          "updatedAfter": {
            "description": "The comparison with current date is strict : updatedAt > updatedAfter",
            "format": "date-time",
            "type": "string"
          },
          "updatedBefore": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "SearchFlowParams": {
        "properties": {
          "cursor": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Cursor"
              }
            ],
            "description": "If cursor is provided, then start search starting from that cursor"
          },
          "limit": {
            "$ref": "#/components/schemas/Limit"
          },
          "where": {
            "$ref": "#/components/schemas/SearchFlowFilters"
          }
        },
        "required": [
          "where"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Use authorization code to access data of another user after he gives his consent.\nUse client credentials to access your own data (similar to api key).\n\nWhen initiating the Authorization Code Flow, you can pre-fill specific user and company details by appending the following query parameters to the `/oauth2/authorize` endpoint:\n- `login_hint`: Prefills the user's email address.\n- `superpdp_company_number`: Prefills the identification number of the company. Required if `superpdp_company_number_scheme` is provided\n- `superpdp_company_number_scheme`: one of `sandbox`, `fr_siren`, `be_numero_entreprise`. Required if  `superpdp_company_number` is provided",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/oauth2/authorize",
            "scopes": {},
            "tokenUrl": "/oauth2/token"
          },
          "clientCredentials": {
            "scopes": {},
            "tokenUrl": "/oauth2/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "name": "SUPER PDP"
    },
...
    "title": "AFNOR Flow Service",
    "version": "1.3.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/flows/search": {
      "post": {
        "description": "Retrieves a set of flows matching the provided search criteria:\n  - Need at least one criterion to be specified\n  - Assuming a logical AND when combining criteria\n  - Assuming a logical OR for criteria allowing a list of values\n\nPagination works with 2 different ways:\n  - with dates and the updatedAfter property\n    - the comparison with current date is strict : updatedAt > updatedAfter\n  - with cursors (more secure)\n    - in case of pagination a nextCursor will be returned\n    - next pagination call shall set cursor value to nextCursor value of last call\n",
        "operationId": "searchFlows",
        "parameters": [
          {
            "description": "Header parameter used to correlate logs from several components",
            "in": "header",
            "name": "Request-Id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The organization that is aimed in a multi tenancy context",
            "in": "header",
            "name": "Organization-Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/FlowSearchRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/FlowSearchResponse"
          },
          "400": {
            "$ref": "#/components/responses/Error400Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401Response"
          },
          "403": {
            "$ref": "#/components/responses/Error403Response"
          },
          "429": {
            "$ref": "#/components/responses/Error429Response"
          },
          "500": {
            "$ref": "#/components/responses/Error500Response"
          },
          "503": {
            "$ref": "#/components/responses/Error503Response"
          }
        },
        "summary": "Select flows upon criteria",
        "tags": [
          "Flow"
        ]
      }
    },
    "/v1/flows/{flowId}": {
      "get": {
        "description": "Download a file related to a given flow:\n  - an invoice\n  - a life cycle\n  - an e-reporting\n",
        "operationId": "getFlow",
        "parameters": [
          {
            "description": "Header parameter used to correlate logs from several components",
            "in": "header",
            "name": "Request-Id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The organization that is aimed in a multi tenancy context",
            "in": "header",
            "name": "Organization-Id",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Flow identifier",
            "in": "path",
            "name": "flowId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NotOnlyUuid"
            }
          },
          {
            "description": "This parameter allows to provide the type of file to be downloaded, can be either one:\n  - Metadata [Default]: provides the flow metadata as a JSON payload, no download\n  - Original: the document that has been initially sent/provided by the emitter\n  - Converted: the document that has been optionally converted by the system\n  - ReadableView: the document that has been optionally generated as the readable file\n",
            "in": "query",
            "name": "docType",
            "schema": {
              "default": "Metadata",
              "enum": [
                "Metadata",
                "Original",
                "Converted",
                "ReadableView"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/FlowGetResponse"
          },
          "400": {
            "$ref": "#/components/responses/Error400Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401Response"
          },
          "403": {
            "$ref": "#/components/responses/Error403Response"
          },
          "404": {
            "$ref": "#/components/responses/Error404Response"
          },
          "429": {
            "$ref": "#/components/responses/Error429Response"
          },
          "500": {
            "$ref": "#/components/responses/Error500Response"
          },
          "503": {
            "$ref": "#/components/responses/Error503Response"
          }
        },
        "summary": "Download the file of a flow",
        "tags": [
          "Flow"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://api.superpdp.tech/afnor-flow"
    }
  ],
...
}